backend-rk-main/requirements.txt
Flask
flask-cors
llama-index-llms-openai
python-dotenv
asyncpg
